@@ -12,7 +12,7 @@ import org.springframework.web.bind.annotation.*;
 import org.springframework.validation.Errors;
 
 
-import javax.validation.Valid;
+import jakarta.validation.Valid;
 
 @RestController
 @AllArgsConstructor
